TrAISformer/pyproject.toml
[tool.poetry]
name = "traisformer"
version = "0.1.0"
description = ""
authors = ["Ukhatov Alexander <alexander.ukhatov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
matplotlib = "^3.8.2"
numpy = "^1.26.3"
tqdm = "^4.66.1"
ipykernel = "^6.29.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
